Using Microsoft.Extension.Logging and DependencyInjection

+++ FileSync/FileSynchronizer.cs
@@ -1,35 +1,32 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Reactive;
 using System.Reactive.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using FileSync.Comparers;
 using FileSync.FileWatchers;
 using FileSync.Operations;
-using Serilog;
+using Microsoft.Extensions.Logging;
 
 namespace FileSync
 {
     public class FileSynchronizer
     {
         private readonly IAppConfig _appConfig;
-        private readonly ILogger _logger;
+        private readonly string _dest;
         private readonly IDirectoryStructureComparer _directoryStructureComparer;
         private readonly IFileComparer _fileComparer;
         private readonly IFileCopier _fileCopier;
         private readonly IFileDeleter _fileDeleter;
         private readonly IFileMerger _fileMerger;
         private readonly IFileWatcher _fileWatcher;
+        private readonly ILogger<FileSynchronizer> _logger;
 
         private readonly string _src;
-        private readonly string _dest;
 
         public FileSynchronizer(
             IAppConfig appConfig,
-            ILogger logger,
+            ILogger<FileSynchronizer> logger,
             IDirectoryStructureComparer directoryStructureComparer,
             IFileComparer fileComparer,
             IFileCopier fileCopier,
@@ -52,7 +49,7 @@ namespace FileSync
 
         public void Sync()
         {
-            _logger.Information($"Synchronizing from {_src} to {_dest}...");
+            _logger.LogInformation($"Synchronizing from {_src} to {_dest}...");
 
             EnsureSrcAndDest();
 
@@ -62,15 +59,15 @@ namespace FileSync
             {
                 Parallel.ForEach(pairs, new ParallelOptions {MaxDegreeOfParallelism = 4}, SyncOnPair);
             }
-            catch (Exception e)
+            catch
             {
-                _logger.Verbose("Synced files = {" + string.Join(", ", pairs.Where(p => p.HasSynced).Select(p => p.SourcePath)) + "}");
-                _logger.Verbose("Not synced files = {" + string.Join(", ", pairs.Where(p => !p.HasSynced).Select(p => p.SourcePath)) + "}");
+                _logger.LogDebug("Synced files = {" + string.Join(", ", pairs.Where(p => p.HasSynced).Select(p => p.SourcePath)) + "}");
+                _logger.LogDebug("Not synced files = {" + string.Join(", ", pairs.Where(p => !p.HasSynced).Select(p => p.SourcePath)) + "}");
 
                 throw;
             }
 
-            _logger.Information("Synchronized.");
+            _logger.LogInformation("Synchronized.");
         }
 
         private void SyncOnPair(Pair pair)
@@ -85,15 +82,11 @@ namespace FileSync
 
             // Src file does not exist, delete the file in the dest directory
             if (isEmptySrcFile)
-            {
                 Delete(destFileRelativePath);
-            }
 
             // Dest file does not exist, copy the src file to dest
             else if (isEmptyDestFile)
-            {
                 Copy(srcFileRelativePath);
-            }
 
             // If both src and dest files exist, compare and sync them
             else
@@ -112,10 +105,7 @@ namespace FileSync
 
         private void Delete(string destFileRelativePath)
         {
-            if (!_appConfig.KeepRemovedFilesInDest)
-            {
-                _fileDeleter.Delete(Path.Combine(_dest, destFileRelativePath));
-            }
+            if (!_appConfig.KeepRemovedFilesInDest) _fileDeleter.Delete(Path.Combine(_dest, destFileRelativePath));
         }
 
         private void Copy(string srcFileRelativePath)
@@ -132,10 +122,7 @@ namespace FileSync
             var destFilePath = Path.Combine(_dest, destFileRelativePath);
 
             var isDifferentFile = !_fileComparer.GetIsEqualFile(srcFilePath, destFilePath);
-            if (isDifferentFile)
-            {
-                _fileMerger.Merge(srcFilePath, destFilePath);
-            }
+            if (isDifferentFile) _fileMerger.Merge(srcFilePath, destFilePath);
         }
 
         private void EnsureSrcAndDest()
